AI_For_Ayurveda/requirements.txt
tensorflow>=2.19.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-chroma>=0.0.1
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
sentence-transformers==2.2.2
transformers==4.36.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0 